+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+
+- support for cc, pitch-bend and set-bank-offset
+
 ## [0.2.0] - 2021-12-22
 
 ### Changed
@@ -11,5 +17,6 @@
 
 Initial release.
 
+[Unreleased]: https://github.com/omkamra/fluidsynth/compare/0.2.0...HEAD
 [0.2.0]: https://github.com/omkamra/fluidsynth/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/omkamra/fluidsynth/tree/0.1.0